+++ yirgacheffe/_layers/__init__.py
@@ -1,6 +1,5 @@
 from .base import YirgacheffeLayer
 from .rasters import RasterLayer, InvalidRasterBand
-from .rescaled import RescaledRasterLayer
 from .reprojected import ReprojectedRasterLayer, ResamplingMethod
 from .vectors import VectorLayer
 from .area import UniformAreaLayer